Some minor fixes
