+++ packages/client/src/reducers/callApi.ts
@@ -1,6 +1,8 @@
 import { Action } from 'redux';
-import { isApiCallAction, IApiCallAction } from 'actions';
+import { isApiCallAction, IApiCallAction } from 'actions/makeApiCall';
 
+// This is the interface each API call will abide by. The 'response'
+// field simply tracks the ID(s) to pull out of the 'entities' store
 export interface IAPIState {
   isInitiallyLoaded: boolean;
   isFetching: boolean;
@@ -14,6 +16,9 @@ export const initialState: IAPIState = {
   response: undefined
 };
 
+// Generic reducer to process a normalized API response. With this, we
+// guarantee a 'loading' state while the API is responding, a consistent
+// response format, and (if implemented) consistent error handling
 const callApi = (
   { request, success, failure }: IApiCallActionTypes
 ) => {
@@ -23,7 +28,7 @@ const callApi = (
   ): IAPIState => {
     const { type } = action;
 
-    if ([request, success, failure].indexOf(type) < 0) {
+    if ([request, success, failure].indexOf(type) < 0 || !isApiCallAction(action)) {
       return state;
     }
 
@@ -32,8 +37,8 @@ const callApi = (
 
     if (type === request) {
       // Maintain the current state and just ensure certain flags are set properly
-      newState.isFetching = false;
-    } else if (isApiCallAction(action) && type === success && action.response) {
+      newState.isFetching = true;
+    } else if (type === success && action.response) {
       newState.isFetching = false;
       newState.errorCode = undefined;
       // Don't store entities since those are stored elsewhere in Redux
